# api/openapi.yaml
openapi: 3.1.0
info:
  title: FinalStore API
  description: |
    REST API for the FinalStore e-commerce platform.
    All authenticated endpoints expect `Authorization: Bearer <jwt>`.
  version: 0.1.0
  contact:
    name: FinalStore
servers:
  - url: http://localhost:5050
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        price: { type: number, format: float }
        stock: { type: integer, minimum: 0 }
        description: { type: string }
        user_id: { type: string }
        created_at: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name: { type: string }
        price: { type: number, format: float }
        stock: { type: integer, minimum: 0 }
        description: { type: string }
...
security:
  - bearerAuth: []
paths:
...
  /signup:
    post:
      tags: [auth]
      security: []
      summary: Create regular user account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '200': { description: created }
        '400': { description: invalid input, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /signupBusiness:
    post:
      tags: [auth]
      security: []
      summary: Create business account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/SignupRequest'
                - type: object
                  properties:
                    company_name: { type: string }
                    company_id: { type: string }
      responses:
        '200': { description: created }
        '400': { description: invalid input }
  /login:
    post:
      tags: [auth]
      security: []
      summary: Authenticate and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { description: invalid credentials }
...
  /products:
    get:
      tags: [products]
      security: []
      summary: List products with optional filters
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [products]
      summary: Create a product (admin/business)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '200':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
...